Renamed and moved Polymorphic sample.
