--- aspnetcore/mvc/advanced/custom-model-binding/3.0sample/PolymorphicModelBinding/Pages/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-@using PolymorphicModelBinding
-@namespace PolymorphicModelBinding.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
