+++ aspnetcore/mvc/advanced/custom-model-binding/samples/3.x/PolymorphicModelBindingSample/Pages/Error.cshtml.cs
@@ -7,7 +7,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.Extensions.Logging;
 
-namespace PolymorphicModelBinding.Pages
+namespace PolymorphicModelBindingSample.Pages
 {
     [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
     public class ErrorModel : PageModel
