+++ aspnetcore/mvc/advanced/custom-model-binding/samples/3.x/PolymorphicModelBindingSample/Pages/Index.cshtml.cs
@@ -6,7 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.Extensions.Logging;
 
-namespace PolymorphicModelBinding.Pages
+namespace PolymorphicModelBindingSample.Pages
 {
     public class IndexModel : PageModel
     {
