+++ aspnetcore/mvc/advanced/custom-model-binding/samples/3.x/PolymorphicModelBindingSample/Pages/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+@using PolymorphicModelBindingSample
+@namespace PolymorphicModelBindingSample.Pages
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
