+++ aspnetcore/mvc/advanced/custom-model-binding/samples/3.x/PolymorphicModelBindingSample/wwwroot/lib/jquery/LICENSE.txt
